pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "paper-context"
version = "0.1.0"
description = "Phase 0 runtime skeleton for the Paper Context MVP."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
  "alembic>=1.16,<2.0",
  "fastapi>=0.115,<1.0",
  "fastmcp>=2.13,<3.0",
  "orjson>=3.10,<4.0",
  "psycopg[binary]>=3.2,<4.0",
  "pydantic-settings>=2.7,<3.0",
  "python-multipart>=0.0.20,<1.0",
  "sqlalchemy>=2.0,<3.0",
  "uvicorn[standard]>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28,<1.0",
  "mypy>=1.15,<2.0",
  "pytest>=8.3,<9.0",
  "ruff>=0.11,<1.0",
]

[project.scripts]
paper-context = "paper_context.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/paper_context"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = false
warn_unused_ignores = true
warn_redundant_casts = true
plugins = []
